// src/types/collections.d.ts
type GenericObject = {
  id: string
  name: string
}

type PriceComparisonType = {
  title: string
  items: GenericObject[]
  source: GenericObject
  target: GenericObject
  id?: string
}

type UserSettings = {
  structures?: GenericObject[]
  priceComparisons?: PriceComparisonType[]
}

export type User = {
  access_token: string
  refresh_token: string
  expiration: number
  name: string
  playerId: number
  parentPlayerId?: number
  settings?: UserSettings
}

export type UserDocument = WithID<User>